Basic working functions
DalekV2MainV2.py will contain the main control code.  

It uses DalekV2Drive.py (which contains the base motor control code for forward, backwards, etc) to allow me to confirm that the base functions work. What it has done is prove that my motors are not up to the job, but that is a different subject.  There are placeholder comments for procedures (abilities) that I need to add, such as line following etc...

Depending on how big the code gets I may break some of it out into other modules.