Added Line following
I have added line following in 2 forms.

One form is to use a webcam and one is to use the pi cam.

The small bit of testing I have done so far implies that the webcam approach may be better (faster) but I am not totally sure now as the PiCam is on a gimbal that I can move to search for lines rather than moving the whole bot.

In addition the actual line following algorythm needa a great deal of further testing.